[Shivaraj] Refactored code with Pytest and adding note and choice with authentication
